allow algo.find_duplicate_files to skip paths it's already seen
because it is, of course, possible to invoke finddupes with the same dir
twice, or a dir and a subdir of it. this will prevent finddupes from
reporting the name multiple times.

however this behavior needs to be optional as correlate.py will need to
handle this case by *not* de-duping paths. as of now, when it comes
time to compare files by content, it will do an unnecessary compare of
the file's bytes against itself, but that's an optimization that can be
made later rather than a conceptual wrinkle that needs to be sorted out
right now