gentle start to a refactor to make serialization a bit more coherent
the problem is marking files. conceptually this is just a boolean that
should be attached to otherwise immutable-by-intent objects
representing file contents. some code will easier to follow if we
just flip a boolean property rather than relying on membership in a
list. report parsing stands to benefit, and report formatting will be
movable to near the parsing code