reverse direction a bit
i think there's a way we can take a set-membership-based approach to
marking files that won't be too crazy.  it still won't be as 'normal' as
trying to craft a .marked property, but i think it still results in an
acceptable API and keeps class explosions to a minimum. and we don't
have to thread a precarious mix of mutability and immutability
everywhere