report now a bit more robust, round-trippable, and self-contained
the format has changed so that blank lines no longer delimit dupe-sets.
instead, sets are headed by the keyword "set". instances are headed
by the word "instance", after which every path is understood to point to
the same file, or the word "singletons", after which every path is
understood to be a separate file. if neither is specified and paths are
specified under just the word "set", "singletons" is assumed.

comments are permitted on keyword lines after a #. paths are serialized
using repr, and preference marking is done with leading ascii
punctuation that is not ' " or #.

instances are now deserialized as well, mostly so that the same data
structures can be used when searching and when re-reading reports.
however any assumptions about hard-links should probably be verified
against the file system